refactor(bench): split hotspot summary helpers
